mini-projet/person-service/requirements.txt
flask
flask_sqlalchemy
flask_cors
pyjwt
